Add db integration

+++ models.py
@@ -12,45 +12,32 @@ class User(Base):
     chat_id = sq.Column(sq.Integer)
 
     def __str__(self):
-        return str(self.chat_id)
+        return f"User: id = {self.id}, chat_id = {self.chat_id}"
 
 class UserCard(Base):
     __tablename__ = "cards"
 
-    id = sq.Column(sq.Integer, primary_key=True)
-    card = sq.Column(sq.String)
-    meaning = sq.Column(sq.String)
+    id = sq.Column(sq.Integer, primary_key=True, autoincrement=True)
+    card_name = sq.Column(sq.String)
+    user_id = sq.Column(sq.Integer, sq.ForeignKey("users.id"), nullable=False)
+
+    user = relationship("User", backref="cards")
 
     def __str__(self):
-        return f"{self.id}: {self.word} - {self.translation}"
+        return f"UserCard: id = {self.id}, key = {self.card_name}, user_id = {self.user_id}"
 
-class UserMeaning(Base):
+class UserCardMeaning(Base):
     __tablename__ = "meanings"
 
     id = sq.Column(sq.Integer, primary_key=True)
-    user_id = sq.Column(sq.Integer, sq.ForeignKey("users.id"))
-    card_id = sq.Column(sq.Integer, sq.ForeignKey("cards.id"))
-    meaning = sq.Column(sq.String)
+    card_meaning = sq.Column(sq.String)
+    card_id = sq.Column(sq.Integer, sq.ForeignKey("cards.id"), nullable=False)
 
-    user = relationship(User, backref="users")
-    word = relationship(UserCard, backref="cards")
+    card = relationship("UserCard", backref="meanings")
 
     def __str__(self):
-        return f"{self.id}: {self.user_id} - {self.word_id}"
+        return f"UserCardMeaning: id = {self.id}, meaning = {self.card_meaning}, card_id = {self.card_id}"
 
-
-
-def add_user(session, user_id):
-    pass
-
-def add_card(session, user_id, card):
-    pass
-
-def add_card_meaning(session, card_id, meaning):
-    pass
-
-def get_cards(session, user_id):
-    pass
 
 
 def create_engine(DSN):
@@ -59,6 +46,7 @@ def create_engine(DSN):
 def create_session_maker(engine):
     return sessionmaker(bind=engine)
 
-def create_tables(engine):
-    Base.metadata.drop_all(engine)
+def create_tables(engine, drop_all=False):
+    if drop_all:
+        Base.metadata.drop_all(engine)
     Base.metadata.create_all(engine)
